+++ java-io-base/src/test/java/_test/io/ResourceId.java
@@ -2,14 +2,15 @@ package _test.io;
 
 import internal.io.text.InternalTextResource;
 import lombok.NonNull;
-import nbbrd.io.Resource;
-import nbbrd.io.text.TextResource;
 
 import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+
+import static nbbrd.io.Resource.newInputStream;
+import static nbbrd.io.text.TextResource.newBufferedReader;
 
 @lombok.Value
 public class ResourceId {
@@ -21,7 +22,7 @@ public class ResourceId {
     String name;
 
     public InputStream open() throws IOException {
-        return Resource.getResourceAsStream(anchor, name).orElseThrow(IOException::new);
+        return newInputStream(anchor, name);
     }
 
     public Path copyTo(Path temp) throws IOException {
@@ -45,7 +46,7 @@ public class ResourceId {
     }
 
     public BufferedReader open(Charset encoding) throws IOException {
-        return TextResource.getResourceAsBufferedReader(anchor, name, encoding).orElseThrow(IOException::new);
+        return newBufferedReader(anchor, name, encoding.newDecoder());
     }
 
     public String copyToString(Charset encoding) throws IOException {